+++ src/foreign/alloc/collections/mod.rs
@@ -1,0 +1,5 @@
+mod binary_heap;
+mod btree_map;
+mod btree_set;
+mod linked_list;
+mod vec_deque;
