+++ src/foreign/core/bool.rs
@@ -1,0 +1,12 @@
+use crate::{Arbitrary, Result, Unstructured};
+
+impl<'a> Arbitrary<'a> for bool {
+    fn arbitrary(u: &mut Unstructured<'a>) -> Result<Self> {
+        Ok(<u8 as Arbitrary<'a>>::arbitrary(u)? & 1 == 1)
+    }
+
+    #[inline]
+    fn size_hint(depth: usize) -> (usize, Option<usize>) {
+        <u8 as Arbitrary<'a>>::size_hint(depth)
+    }
+}
