+++ src/foreign/core/mod.rs
@@ -1,0 +1,20 @@
+//! Implementations of [`Arbitrary`] for [`core`] types.
+//!
+//! [`Arbitrary`]: crate::Arbitrary
+
+mod array;
+mod bool;
+mod cell;
+mod char;
+mod iter;
+mod marker;
+mod num;
+mod ops;
+mod option;
+mod result;
+mod slice;
+mod str;
+mod sync;
+mod time;
+mod tuple;
+mod unit;
