+++ src/foreign/core/sync/mod.rs
@@ -1,0 +1,1 @@
+mod atomic;
