+++ src/foreign/mod.rs
@@ -1,0 +1,7 @@
+//! Implementations of [`Arbitrary`] for foreign types.
+//!
+//! [`Arbitrary`]: crate::Arbitrary
+
+mod alloc;
+mod core;
+mod std;
