+++ src/foreign/std/collections/mod.rs
@@ -1,0 +1,2 @@
+mod hash_map;
+mod hash_set;
