+++ src/foreign/std/ffi/c_str.rs
@@ -1,0 +1,5 @@
+// impl Arbitrary for Box<CStr> {
+//     fn arbitrary(u: &mut Unstructured<'_>) -> Result<Self> {
+//         <CString as Arbitrary>::arbitrary(u).map(|x| x.into_boxed_c_str())
+//     }
+// }
